Fixed errors in patient flags module.
