+++ omod/src/main/java/org/openmrs/module/webservices/rest/web/controller/PatientFlagRestController.java
@@ -11,7 +11,7 @@
  *
  * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
  */
-package org.openmrs.module.patientflag_rest.web.controller;
+package org.openmrs.module.webservices.rest.web.controller;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
