+++ omod/src/main/java/org/openmrs/module/webservices/rest/web/wrapper/FlagWrapper.java
@@ -1,14 +1,25 @@
-package org.openmrs.module.patientflag_rest.web.wrapper;
+/**
+ * The contents of this file are subject to the OpenMRS Public License
+ * Version 1.0 (the "License"); you may not use this file except in
+ * compliance with the License. You may obtain a copy of the License at
+ * http://license.openmrs.org
+ *
+ * Software distributed under the License is distributed on an "AS IS"
+ * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
+ * License for the specific language governing rights and limitations
+ * under the License.
+ *
+ * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
+ */
+package org.openmrs.module.webservices.rest.web.wrapper;
 
 import java.io.Serializable;
 import java.util.List;
 
 import org.openmrs.BaseOpenmrsData;
 
-/**
- * @author Stélio Moiane
- *
- */
+
+
 public class FlagWrapper extends BaseOpenmrsData implements Serializable {
 
 	/**
