+++ omod/src/main/webapp/manage.jsp
@@ -1,4 +1,3 @@
-<%@ include file="/WEB-INF/template/include.jsp"%>
 <%@ include file="/WEB-INF/template/header.jsp"%>
 
 <%@ include file="template/localHeader.jsp"%>
