+++ omod/src/main/webapp/template/localHeader.jsp
@@ -1,4 +1,3 @@
-<spring:htmlEscape defaultHtmlEscape="true" />
 <ul id="menu">
 	<li class="first"><a
 		href="${pageContext.request.contextPath}/admin"><spring:message
